fix: github pr comments
